+++ lib/msal-common/src/constants/AADServerParamKeys.ts
@@ -58,3 +58,4 @@ export const DOMAIN_HINT = "domain_hint";
 export const X_CLIENT_EXTRA_SKU = "x-client-xtra-sku";
 export const BROKER_CLIENT_ID = "brk_client_id";
 export const BROKER_REDIRECT_URI = "brk_redirect_uri";
+export const INSTANCE_AWARE = "instance_aware";