+++ lib/msal-angular/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @azure/msal-angular
 
-<!-- This log was last generated on Tue, 18 Feb 2025 18:51:41 GMT and should not be manually modified. -->
+<!-- This log was last generated on Tue, 11 Mar 2025 18:51:25 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 4.0.6
+
+Tue, 11 Mar 2025 18:51:25 GMT
+
+### Patches
+
+- Bump @azure/msal-browser to v4.7.0
+- Bump eslint-config-msal to v0.0.0
 
 ## 4.0.5